Tyding Up

+++ src/app/messages/message/message.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from 'src/app/_models/message';
 import { Pagination } from 'src/app/_models/pagination';
+import { ConfirmWindowService } from 'src/app/_services/confirm-window.service';
 import { MessageService } from 'src/app/_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessageComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageSerive: MessageService) { }
+  constructor(private messageSerive: MessageService, private confirmService: ConfirmWindowService) { }
 
   ngOnInit(): void {
     this.loadMessages()
@@ -41,9 +42,14 @@ export class MessageComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageSerive.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result=> {
+      if(result){
+        this.messageSerive.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+        })
+      }
     })
+   
   }
 
 }
